backend/requirements.txt
fastapi
uvicorn[standard]
pymongo
python-dotenv==1.0.1
